Use custom interface for database instead of generic one
In order to support the generic helper functions for transactions and usual
connections, we used to use the generic interface and do manual type-checks
and conversions.

However, as sqlx.DB and sqlx.Tx share the same method signatures, I created a
new interface that contains all used methods and was able to replace the usage
of the generic interface{} with the new one.